feat(data-provider): cast contract return values to appropriate types in HyperIndex and MorphoApi

+++ apps/data-providers/src/hyperIndex/index.ts
@@ -333,7 +333,7 @@ export class HyperIndexDataProvider implements DataProvider {
               abi: oracleAbi,
               functionName: "price",
             });
-            oraclePrices.set(oracle, price);
+            oraclePrices.set(oracle, price as bigint);
           } catch {
             oraclePrices.set(oracle, undefined);
           }
